requirements.txt
flask
gunicorn
moviepy
requests
openai
python-dotenv